first commit
Created a tool to calculate profit and loss using the start of year basis valuation for
all trading accounts and all trades for that year. The script will match each sell trade
to a buy trade, use the minimum size for each find the profit and loss and enter it into
a profit and loss table. Once all sells have been processesed the remaining buys will be
considered the basis for the end of the year. Both a profit and loss and basis csv will
result from running the script.

Also, wash trades, recorded losses that are purchased back in a period of less than
30 days, are tracked and accounted for in the profit and loss table. Each basis entry
that results in a wash trade will also record the adjusted basis for the unacounted for
loss.