Created exchange_api class and tests
After accepting the flaws in this script, I have surrendered to the fact
that tests and refactor are needed. Not knowing where to start, I picked
the calls to the exchange api. I did this because for one, they are
simple and for two, I was interested in writing tests that did not need
to actually call the api. This commit removes the api call handling to a
sperate package and adds testing for all functionality.