Added wash loss logic to ProfitAndLoss
Wash logic made sense to add in the ProfitAndLoss class since there was
already sizing logic that could be used. By passing the basis trade into
the wash_loss method, the trade can be checked for size in the current
trading asset. This is not straight forward with btc quote pair sells
that end up basis trades for btc since the typically sizes are grabed
directly for the base pair, but for a quote pair they need to be
calculated based on size, price, and fee. This logic was already added
to the ProfitAndLoss class previously to grab sizes for basis and
proceeds trades and make sure they matched. Wash trade logic in the
trade processor is mid flight. Still need to handle mismatched sizes for
wash trades and add functionality to reset it when the most recent
potential wash loss trade is older than thirty days.