Finished wiring in the trade_processor
Processor now is now operating in the main flow. Deleted old code. There
is some additional rounding needed for BTC, likely for trades with BTC
as a quote pair since all other assets are not having issues. Also would
like to concatenate two dataframes that have the same headers into one
and print it to csv, but how to do that is not immediately obvious with
the matching headers.