Fixed issue with wash sale before loss
Losses were being transfered to basis that occured earlier in time, but
if that basis had already sold, the profit and loss of that sale were
not adjsuted. Also a simple buy and sell at a loss would get washed
agasinst itself and the loss vanishes. Now washing a loss can not occur
against the basis trade belonging to the loss and a wash trades that
have already sold will get the basis adjusted and the profit and loss
adjusted.