Issue #16 Refactored Input Code
Created tests and restructured method of checking for usd per btc.
Datetime formats were not specified in DataFrame.to_csv which led to an
output format that was inconsistent to the input format. If you were to
process the same input files twice for some reason, and the first time
you didn't have usd per btc values, the code now re writes the input
files with those values. If the times are not consistent, then the
second run breaks reading the time format. This will also be required
later for Issue #22. Close #16.