GH-32 Added Sorter for loading data and arranging by datetime
Prior, processing was done one asset at a time. Due to complexity of
managing transferring of assets, it was difficult to think of a way to
handle one asset at a time without knowing how when assets were going to
be transferred and needing to know what the basis for those assets were
at the time of transfer. By handling all events chronologically, any
transfer will be able to include the basis of the amount of the asset
getting transferred.