GH-32 Added general handler that delegates to each buckets handlers
Also removed ProfitAndLoss object from Entry and created some hacks to
get everything to work. One exception is wash trading. Wash trade code
paths were dependent on the ProfitAndLoss from being accessible on the
Entry rather than creating one using a static function as I hacked it
in. The ProfitAndLoss object should be refactored, as it has become
unwieldy.

the association between entry and profit and loss is such that they
could be combined, but there is also a lot of wash logic in the profit
and loss object that should be extracted out somewhere else. A mapping
from entry to p_l allows for moving forward with the task at hand so
this can be revisited in a different effort.